Make opamp write "O" instead of "gnd" nodes
When opamp writes out a spice netlist, write "0" whenever the node name
is "gnd".

Xyce does not recognize "gnd" as an alias for ground (neither does
SPICE3F5, on which Xyce's netlist language was based).  It does have a
netlist option intended to make it recognize that alias (and others),
but it has a bug that makes it not work when the alias is used inside an
expression, as it is when Qucs-S generates the ideal opamp model as a B
source.

Since most other devices go through pains to convert "gnd" in their node
names into "0", I inserted similar code into the opamp.cpp
"spice_netlist" function.

With this change, schematics created by qucs-sactivefilter now can run
in Xyce (when the "ac simulation" object is replaced by an appropriate
"xyce script" object).